requirements.txt
celery==5.5.1
django==5.1.8
django-celery-beat==2.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf_spectacular==0.28.0
mysqlclient==2.2.7
redis==5.2.1
